some more changes to PollingRedeployer

+++ vertx-testsuite/src/test/java/org/vertx/java/tests/core/redeploy/DefaultRedeployerTest.java
@@ -4,7 +4,11 @@ import org.junit.After;
 import org.junit.AfterClass;
 import org.junit.Before;
 import org.junit.BeforeClass;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TestWatcher;
+import org.junit.rules.Timeout;
+import org.junit.runner.Description;
 
 import static org.junit.Assert.*;
 
@@ -34,7 +38,19 @@ import java.util.concurrent.TimeUnit;
 public class DefaultRedeployerTest {
 
   private static final Logger log = LoggerFactory.getLogger(DefaultRedeployerTest.class);
-  private static final int SLEEP = 1000;
+
+  protected static long SLEEP = 500;
+  
+  // Useful for tests NOT executed via gradle. Gradle hooks into the junit test lifecycle and nicely indents the output.
+  // See http://www.gradle.org/docs/current/dsl/org.gradle.api.tasks.testing.Test.html
+  @Rule public TestWatcher watchman = new TestWatcher() {
+    public void starting(Description descr) {
+      log.info(">> " + descr.getMethodName());
+    }
+  };  
+
+  // Default timeout for all test cases
+  @Rule public Timeout timeout = new Timeout(10_000);
   
   protected static VertxInternal vertx;
   protected TestReloader reloader;
@@ -76,7 +92,9 @@ public class DefaultRedeployerTest {
     while(true) {
 	    Thread.sleep(200);
 	  	try {
-	  		vertx.fileSystem().deleteSync(modRoot.getAbsolutePath(), true);
+	  		if (modRoot.exists()) {
+	  			vertx.fileSystem().deleteSync(modRoot.getAbsolutePath(), true);
+	  		}
 	  		break;
 	  	} catch (DirectoryNotEmptyException ex) {
 	  		if (++count > 20) {
@@ -107,7 +125,7 @@ public class DefaultRedeployerTest {
     Deployment dep = createDeployment("dep1", "my-mod", null);
     red.moduleDeployed(dep);
     Thread.sleep(SLEEP);
-    modifyFile(modDir, "blah.txt");
+    modifyFile(modDir, "foo.js");
     waitReload(dep);
   }
 
@@ -152,7 +170,6 @@ public class DefaultRedeployerTest {
 
   @Test
   public void testDeleteFileInSubDirectory() throws Exception {
-  	log.error("testDeleteFileInSubDirectory");
     String modName = "my-mod";
     File modDir = createModDir(modName);
     createFile(modDir, "foo.js", TestUtils.randomAlphaString(1000));
@@ -161,7 +178,6 @@ public class DefaultRedeployerTest {
     Deployment dep = createDeployment("dep1", "my-mod", null);
     red.moduleDeployed(dep);
     Thread.sleep(SLEEP);
-  	log.error("delete the file 'bar.txt'");
     deleteFile(subDir, "bar.txt");
     waitReload(dep);
   }
@@ -216,12 +232,14 @@ public class DefaultRedeployerTest {
   private File createModDir(String modName) {
     File modDir = new File(modRoot, modName);
     modDir.mkdir();
+    log.info("Create Directory: " + modDir.getAbsolutePath() + "  " + modDir.lastModified());
     return modDir;
   }
 
   private void createFile(File dir, String fileName, String content) throws Exception {
     File f = new File(dir, fileName);
     vertx.fileSystem().writeFileSync(f.getAbsolutePath(), new Buffer(content));
+    log.info("Create File: " + f.getAbsolutePath() + "  " + f.lastModified());
   }
 
   private void modifyFile(File dir, String fileName) throws Exception {
@@ -229,16 +247,19 @@ public class DefaultRedeployerTest {
     FileWriter fw = new FileWriter(f, true);
     fw.write(TestUtils.randomAlphaString(500));
     fw.close();
+    log.info("Modify File: " + f.getAbsolutePath() + "  " + f.lastModified());
   }
 
   private void deleteFile(File dir, String fileName) throws Exception {
     File f = new File(dir, fileName);
+    log.info("Delete File: " + f.getAbsolutePath());
     f.delete();
   }
 
   private File createDirectory(File dir, String dirName) throws Exception {
     File f = new File(dir, dirName);
     vertx.fileSystem().mkdirSync(f.getAbsolutePath());
+    log.info("Create Directory: " + f.getAbsolutePath() + "  " + f.lastModified());
     return f;
   }
 
@@ -273,9 +294,9 @@ public class DefaultRedeployerTest {
     }
 
     private void checkDeps(Set<Deployment> deps) {
-      assertEquals(deps.size(), reloaded.size());
+      assertEquals("Number of expected Deployments does not match: ", deps.size(), reloaded.size());
       for (Deployment dep: deps) {
-        assertTrue(reloaded.contains(dep));
+        assertTrue("Expected the Deployment was reloaded: " + dep, reloaded.contains(dep));
       }
     }
   }
